[IMP] account_payment_kmitl: merge account_payment_budget module
Consolidate budget integration for payments into account_payment_kmitl,
eliminating the separate account_payment_budget module.

- Move budget validation (_check_analytic_distribution_complete) and
  commitment consumption (_post) to account_move_kmitl so budget logic
  lives at the journal entry level
- Remove budget.commitment.mixin from account.payment (already on
  account.move, accessible via _inherits)
- Keep _onchange_budget_commitment_id on payment (required because
  _inherits does not cascade onchange handlers)
- Merge Budget tab XML, security rules, and budget dependency into
  account_payment_kmitl
- Remove account_payment_budget from kmitl_depends_dev
- Delete account_payment_budget module
