--- account_payment_budget/models/__init__.py
@@ -1,5 +1,0 @@
-# -*- coding: utf-8 -*-
-# Copyright (C) 2024 KMITL
-# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
-
-from . import account_payment
