+++ account_payment_kmitl/__manifest__.py
@@ -12,6 +12,7 @@
         "account",
         "account_move_kmitl",
         "account_fiscal_year_enhance",
+        "budget",
         "l10n_th_bank_payment_export",
         "l10n_th_bank_payment_export_format",
         "thai_date_utils",
