+++ account_payment_kmitl/views/account_payment_views.xml
@@ -44,6 +44,27 @@
             <xpath expr="//field[@name='payment_type']" position="after">
                 <field name="kmitl_payment_type_id" domain="[('direction', '=', payment_type)]"/></xpath>
 
+            <!-- Add Budget tab -->
+            <xpath expr="//sheet" position="inside">
+                <notebook>
+                    <page string="Budget" name="budget" attrs="{'invisible': [('payment_type', '!=', 'outbound')]}">
+                        <group string="Budget">
+                            <group name="budget_info">
+                                <field name="budget_commitment_id" attrs="{'readonly': [('state', '!=', 'draft')]}" />
+                                <field name="analytic_distribution" widget="analytic_distribution" readonly="1"/>
+                            </group>
+                            <group name="analytic_dimensions">
+                                <field name="activity_analytic_id" domain="[('plan_id.code', '=', 'activities')]" options="{'no_create': True}" required="1" attrs="{'readonly': ['|', ('budget_commitment_id', '!=', False), ('state', '!=', 'draft')]}"/>
+                                <field name="department_analytic_id" domain="[('plan_id.code', '=', 'departments')]" options="{'no_create': True}" required="1" attrs="{'readonly': ['|', ('budget_commitment_id', '!=', False), ('state', '!=', 'draft')]}"/>
+                                <field name="fund_analytic_id" domain="[('plan_id.code', '=', 'funds')]" options="{'no_create': True}" required="1" attrs="{'readonly': ['|', ('budget_commitment_id', '!=', False), ('state', '!=', 'draft')]}"/>
+                                <field name="source_analytic_id" domain="[('plan_id.code', '=', 'sources')]" options="{'no_create': True}" required="1" attrs="{'readonly': ['|', ('budget_commitment_id', '!=', False), ('state', '!=', 'draft')]}"/>
+                                <field name="budget_account_id" options="{'no_create': True}" required="1" attrs="{'readonly': ['|', ('budget_commitment_id', '!=', False), ('state', '!=', 'draft')]}"/>
+                            </group>
+                        </group>
+                    </page>
+                </notebook>
+            </xpath>
+
         </field>
     </record>
